Split PlayerQuestionState into 3 classes to avoid a code smell. These classes include PlayerNameState, PlayerDescriptionState, and PlayerSkillTypeState. Also split PlayerCreatorInputValidator into 4 classes to correspond with these changes, including PlayerNameValidator, PlayerDescriptionValidator, PlayerSkillTypeValidator, and PlayerConfirmValidator. Added testing for the new classes and updated the PlayerCreatorManager and corresponding factories to function with the change. Deleted PlayerCreatorInputValidator and PlayerQuestionState.
